requirements.txt.txt
streamlit==1.32.0
pandas==2.2.2
matplotlib==3.8.4
python-dotenv==1.0.1
fastmcp==0.1.6
google-search-results==2.4.2